Move event pipe context and callback into eventtrace
