+++ src/coreclr/nativeaot/Runtime/eventtrace_context.h
@@ -1,0 +1,47 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#ifndef EVENTTRACE_CONTEXT_H
+#define EVENTTRACE_CONTEXT_H
+
+#ifdef FEATURE_EVENT_TRACE
+
+#include "eventtrace_etw.h"
+
+#if !defined(EVENTPIPE_TRACE_CONTEXT_DEF)
+#define EVENTPIPE_TRACE_CONTEXT_DEF
+typedef struct _EVENTPIPE_TRACE_CONTEXT
+{
+    const WCHAR * Name;
+    unsigned char  Level;
+    bool IsEnabled;
+    ULONGLONG EnabledKeywordsBitmask;
+} EVENTPIPE_TRACE_CONTEXT, *PEVENTPIPE_TRACE_CONTEXT;
+#endif // EVENTPIPE_TRACE_CONTEXT_DEF
+
+#if !defined(DOTNET_TRACE_CONTEXT_DEF)
+#define DOTNET_TRACE_CONTEXT_DEF
+typedef struct _DOTNET_TRACE_CONTEXT
+{
+#ifdef FEATURE_ETW
+    RH_ETW_CONTEXT* EtwProvider;
+#endif
+    EVENTPIPE_TRACE_CONTEXT EventPipeProvider;
+} DOTNET_TRACE_CONTEXT, *PDOTNET_TRACE_CONTEXT;
+#endif // DOTNET_TRACE_CONTEXT_DEF
+
+extern "C" DOTNET_TRACE_CONTEXT MICROSOFT_WINDOWS_DOTNETRUNTIME_PROVIDER_DOTNET_Context;
+
+struct _EventFilterDescriptor;
+typedef struct _EventFilterDescriptor EventFilterDescriptor;
+void EventPipeEtwCallbackDotNETRuntime(
+    _In_ GUID * SourceId,
+    _In_ ULONG ControlCode,
+    _In_ unsigned char Level,
+    _In_ ULONGLONG MatchAnyKeyword,
+    _In_ ULONGLONG MatchAllKeyword,
+    _In_opt_ EventFilterDescriptor* FilterData,
+    _Inout_opt_ PVOID CallbackContext);
+
+#endif // FEATURE_EVENT_TRACE
+#endif // EVENTTRACE_CONTEXT_H
